Fix/Feat: Fix migration error, changing time column type in db, adding a finished game listener

+++ src/models/move.ts
@@ -31,7 +31,7 @@ Move.init({
     san: { type: DataTypes.TEXT, allowNull: false },
     lan: { type: DataTypes.TEXT, allowNull: false },
     promotion: { type: DataTypes.TEXT, allowNull: true },
-    playerTimeLeft: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 600 },
+    playerTimeLeft: { type: DataTypes.DOUBLE, allowNull: false, defaultValue: 600 },
     moverId: { type: DataTypes.UUID, references: { model: 'users', key: 'id' }, allowNull: false },
     moveScore: { type: DataTypes.FLOAT, allowNull: false, defaultValue: 0 }
 }, {
